Improved behavior of CachedFilesMixin.post_process
When using a network storage backend (like S3), post_process can take a long
time, during which no feedback is presented when running collectstatic. All of
the modified files are processed upfront, and there's progress in the console
only when they're all finished.

Changing post_process to a generator makes it easier to print out feedback as
each file is processed.