Use local file to generate cache hash
Removes the need to fetch file out of target storage backend to check
whether a file has changed. If the target storage is a local disk, this
doesn't make much of a difference, but with remote storage (like S3),
this can have a major performance boost.

Additionally, improves feedback during post_process to indicate whether
the hashed files are being copied or not.